+++ HISTORY.rst
@@ -6,6 +6,8 @@ History
 1.13.3 (2014-10-13)
 ~~~~~~~~~~~~~~~~~~~
 
+- Fixing bug: source updates failed when using --locale and --types flags
+  together.
 - Updating bindings version and fixing code accordingly.
 
 1.13.2 (2014-10-05)